- Refactored names to MVP pattern which is more approriate for what is happening.

+++ src/GameEditor/GameEditor.Wcf.Harness/Models/GameModel.cs
@@ -73,6 +73,7 @@ namespace GameEditor.Wcf.Harness.Models
 
         public void DeleteArmyUnitDefinition(int id)
         {
+            if (id == 0) return;
             using (var client = new WarGameServiceClient())
             {
                 client.DeleteArmyUnitDefinition(CurrentArmyDefinitionId, id);
